+++ batma/__init__.py
@@ -19,20 +19,23 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE 
 # SOFTWARE.
 
-'''Batma 2D game engine provides an intuitive structure for game development and 
+'''Batma's main module.
+
+Batma 2D game engine provides an intuitive structure for game development and 
 other graphical applications.
 
-Features includes:
+Main features:
 
-    * Automatic import regular image atlas in a list of images
-    * Animated Sprites creation from image atlas or gifs
-    * Easy-to-use pyglet components, e.g., sprites, texts and resource locations
-    * Input state for keyboard and mouse (also inspired by XNA)
-
+  - Automatic import regular (equally spaced) images atlas in a list of images;
+  - Animated sprites creation from image atlas or gifs;
+  - Easy-to-use pyglet components, e.g. sprites, texts and resource location;
+  - Input state for for keyboard and mouse (similar to XNA);
+  - Primitives drawning: opengl defaults, circle, rectangle;
+  - 2D cameras handling
+  - Scenes based games
 '''
 
-__docformat__ = 'restructuredtext en'
-__version__ = "0.0.1"
+__version__ = "0.1"
 __author__ = "Renato de Pontes Pereira"
 version = __version__
 
@@ -45,14 +48,15 @@ from batma.resource import *
 from batma.sprite import *
 from batma.text import *
 from batma.node import *
-from batma import draw
 from batma.algebra import *
+from batma import draw
 from batma import colors
 import pyglet.window.key as keys
 import pyglet.window.mouse as buttons
 
 def run():
-    '''
-    Starts the game.
+    '''Starts the game.
+
+    It's the same of ``pyglet.app.run()``
     '''
     pyglet.app.run()